Hooked up done/verify tasks buttons in the app. Fixes #63
This can be implemented by including two data attributes on any button eg.

    <button data-task-state="verify" data-task-id="2">Verify Task</button>

There is now a listener in init.js that will update the task id when the
button is clicked. It is up to the view to reflect the change in state.
This listener needs to be moved to a more appropriate location.