Updated convert newlines in Hub and Task descriptions to breaks. Fixes #204
Ended up doing this in the views in the end. Originally I created a tim()
filter to convert all newlines in the template data. However this broke
the forms as textareas ended up with "<br/>" strings in them. So as we only
use this in two places I manually update the descriptions using jQuery.