Fixed bug in patch to Backbone.Events.unbind()
Previous code was assigning the results of list.splice() to the _callbacks
object. [].splice() returns the extracted elements in a new array rather
than the original modified array. Code has been updated to call .splice()
in a separate statement to the assignment. Also we are now cloning the
list variable using [].slice() rather than _.clone().