Fixed issue where tasks would not always render on a hub. Fixes #344
This issue was caused by the tasks on the hub being requested elsewhere
(the dashboard) and the Tasket.getModels() method returning cached stubs
in a collection. This mean the collection would never fire the "refresh"
event (as no request was made) and the UI would not update.

Tasket.getModels() now checks to see if it returning an incomplete
collection and watches loading models for their "change" event. Once the
collection is complete it manually fires the "refresh" event updating
relevant UI.