Hidden dashboard now resets position after resize. Closes #419
This is not particularly smooth at the moment due to the way the dashboard
is hidden with a CSS transform we can't easily update it quickly without it
looking ugly. The other solution would be to revert the dashboard transition,
as this would mean the hidden dashboard always stays in the same place.